backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
boto3==1.34.0
botocore==1.34.0
gunicorn==21.2.0
SQLAlchemy==2.0.22
alembic==1.11.1
PyJWT==2.8.0
passlib==1.7.4

